37-logo-redesign-ai/requirements.txt
flask==2.3.3
pillow==10.0.0
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
scikit-image==0.21.0
matplotlib==3.7.2